% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmediatR.R
\name{bmediatR}
\alias{bmediatR}
\title{Bayesian model selection for mediation analysis function}
\usage{
bmediatR(
  y,
  M,
  X,
  Z = NULL,
  w = NULL,
  kappa = rep(0.001, 8),
  lambda = rep(0.001, 8),
  tau_sq_mu = rep(1000, 8),
  tau_sq_Z = rep(1000, 8),
  phi_sq_X = c(NA, NA, 1, 0.5, NA, 1, NA, 0.5),
  phi_sq_m = c(NA, 1, NA, 0.5, NA, NA, NA, NA),
  phi_sq_y = c(NA, NA, NA, NA, NA, NA, 1, 0.5),
  ln_prior_c = "complete",
  options_X = list(sum_to_zero = T, center = F, scale = F),
  verbose = T
)
}
\arguments{
\item{y}{Vector or single column matrix of an outcome variable. Single outcome variable expected. 
Names or rownames must match across M and X, and Z and w (if provided).}

\item{M}{Vector or matrix of mediator variables. Multiple mediator variables are supported. 
Names or rownames must match across y and X, and Z and w (if provided).}

\item{X}{Design matrix of the driver. Names or rownames must match across y and M, and Z and w (if provided).
One common application is for X to represent genetic information at a QTL as either founder strain haplotypes
in a multiparental population or variant genotypes, though it is generalizable to types of variables.}

\item{Z}{DEFAULT: NULL. Design matrix of covariates that influence the outcome variable. Names or rownames must match across y, M and X, 
and w (if provided).}

\item{w}{DEFAULT: NULL. Vector of weights for individuals in analysis. Names or rownames must match across y, M X, and Z (if provided).
A common use would be for an analysis of strain means, where w would be a vector of the number of individuals per strain.
If no w is given, observations are equally weighted as 1.}
}
\description{
This function takes an outcome, mediator(s), and a driver as a design matrix to perform a Bayesian model selection analysis for mediation.
}
\examples{
bmediatR()
}
