% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmediatR.R
\name{estimate_empirical_prior}
\alias{estimate_empirical_prior}
\title{Calculate empirical priors based on the relationships in the data between X, M, and y}
\usage{
estimate_empirical_prior(ln_prob_data, model = c("complete", "partial"))
}
\arguments{
\item{ln_prob_data}{Marginal likelihoods from bmediatR fit.}

\item{model}{DEFAULT: "complete". If "complete", the marginal likelihoods evaluating the
effects of X on M and Y. If "partial" is specified, the effects of X on M and Y are fixed
as present. A "reactive" mode is currently not supported because of the challenge in
distinguishing the directionality of a relationship.}
}
\description{
This function estimates log prior model probabilities that maximize the marginal likelihoods
of the data from a previous run of bmediatR().
}
\examples{
estimate_empirical_prior()
}
