% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_mediation_bf.R
\name{get_approx_pval}
\alias{get_approx_pval}
\title{Function that calculate approximate p-values for Bayes factors}
\usage{
get_approx_pval(med_bf_object, annot)
}
\description{
This function takes Bayes factor results from mediation_bf() and calculates approximate p-values based
on comparisons to genome results, using the assumption that the vast majority of candidates are not
mediators and can thus be used to characterize a null distribution of Bayes factors.
}
\examples{
get_approx_pval()
}
