% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmediatR.R
\name{model_info}
\alias{model_info}
\title{Definions and encodings for models included in the mediation analysis through Bayesian model selection}
\usage{
model_info()
}
\description{
This function prints out a table describing the models considered in the mediation analysis
}
\examples{
model_info()
}
