% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{model_matrix_from_ID}
\alias{model_matrix_from_ID}
\title{Produce SDP mapping matrix from SDP string}
\usage{
model_matrix_from_ID(M_ID)
}
\arguments{
\item{M_ID}{SDP string. Example is "0,0,0,0,1,1,1,1", which represent two functional 
alleles evenly distributed across the eight founder strains.}
}
\description{
This function takes an SDP string, i.e., "0,0,0,0,1,1,1,1", and converts it into the a
matrix that maps the eight allele haplotype design matrix to < eight functional allele
design matrix.
}
\examples{
model_matrix_from_ID()
}
