% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediation_bf.R
\name{plot_bf}
\alias{plot_bf}
\title{Bayes factor genome scan function}
\usage{
plot_bf(
  med_bf_object,
  bf_type = "lnBF_med",
  med_annot,
  include_chr = c(1:19, "X"),
  expand_lim_factor = 0.025,
  label_thresh = NULL,
  bgcol = "white",
  altcol = "gray",
  altbgcol = "white",
  hlines_col = "gray80",
  col = "black",
  cex = 0.75,
  qtl_dat = NULL,
  outcome_symbol = NULL,
  ...
)
}
\arguments{
\item{med_bf_object}{Output from mediation_bf().}

\item{bf_type}{DEFAULT: "lnBF_med". Bayes factor to be displayed.}

\item{med_annot}{Annotation data for -omic mediators.}

\item{include_chr}{DEFAULT: c(1:19, "X"). Chromosomes to include in plot.}

\item{label_thresh}{DEFAULT: NULL. Label mediators that surpass label_thresh. Default does not add labels.}

\item{qtl_dat}{DEFAULT: NULL. QTL data that includes position of QTL and outcome. Adds ticks to the figure.}

\item{expland_lim_factor}{DEFAULT: 0.025. Scale to increase plot limits by.}
}
\description{
This function takes Bayes factor results from mediation_bf() and plots the genome-wide mediation scan.
}
\examples{
plot_bf()
}
