% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediation_bf.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Posterior probability genome plot function}
\usage{
plot_posterior(
  med_bf_object,
  post_col = c(4, 8),
  med_annot,
  med_var = "protein.id",
  include_chr = c(1:19, "X"),
  expand_lim_factor = 0.025,
  label_thresh = NULL,
  bgcol = "white",
  altcol = "gray",
  altbgcol = "white",
  hlines_col = "gray80",
  col = "black",
  cex = 0.75,
  qtl_dat = NULL,
  outcome_symbol = NULL,
  ...
)
}
\description{
This function takes posterior probability results from mediation_bf() and plots the genome-wide mediation scan.
}
\examples{
plot_posterior()
}
