% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediation.R
\name{plot_posterior_bar}
\alias{plot_posterior_bar}
\title{Barplot of posterior model probabilities}
\usage{
plot_posterior_bar(
  bmediatR_object,
  med_annot = NULL,
  mediator_id,
  med_var = "protein.id",
  stack = FALSE,
  bar_col = c("seagreen4", "seagreen1", "skyblue", "goldenrod1", "goldenrod4", "gray"),
  relabel_x = NULL,
  add_number_labels = FALSE,
  label_size = 5,
  num_dig = 3,
  main = NULL
)
}
\arguments{
\item{bmediatR_object}{Output from bmediatR().}

\item{med_annot}{Annotation data for -omic mediators.}

\item{mediator_id}{Which mediator to plot posterior model probabilities for.}

\item{med_var}{DEFAULT: "protein.id". The column in med_annot to be used as a mediator id.}

\item{stack}{DEFAULT: FALSE. If TRUE, a stacked barplot is produced. If FALSE, a staggered
barplot is produced.}

\item{add_number_labels}{DEFAULT: FALSE. Add posterior probabilities as text above bars.}

\item{label_size}{DEFAULT: 5. Text size of probability labels, when included.}

\item{num_dig}{DEFAULT: 3. The number of digits after the decimal point if probabilities
are being included as labels.}
}
\description{
This function takes posterior probability results from mediation_bf() and plots barplots of posterior model probabilities.
}
\examples{
plot_posterior_bar()
}
