% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediation.R
\name{plot_posterior_odds}
\alias{plot_posterior_odds}
\title{Posterior odds genome plot function}
\usage{
plot_posterior_odds(
  bmediatR_object,
  model_type = c("mediation", "partial", "complete", "colocal"),
  med_annot,
  med_var = "protein.id",
  include_chr = c(1:19, "X"),
  expand_lim_factor = 0.025,
  label_thresh = NULL,
  label_only_chr = NULL,
  bgcol = "white",
  altcol = "gray",
  altbgcol = "white",
  hlines_col = "gray80",
  col = "black",
  cex = 0.75,
  qtl_dat = NULL,
  outcome_symbol = NULL,
  ymax = NULL,
  ymin = NULL,
  ...
)
}
\arguments{
\item{bmediatR_object}{Output from bmediatR().}

\item{model_type}{DEFAULT: "mediation". Specifies which model(s)'s posterior probabilities are to be included in the numerator of the posterior odds and then displayed for
for genome-wide mediators.}

\item{med_annot}{Annotation data for -omic mediators.}

\item{include_chr}{DEFAULT: c(1:19, "X"). Chromosomes to include in plot.}

\item{label_thresh}{DEFAULT: NULL. Label mediators that surpass label_thresh. Default does not add labels.}

\item{label_only_chr}{DEFAULT: NULL. Only label mediators that pass label_thresh on the specified chromosome.}

\item{qtl_dat}{DEFAULT: NULL. QTL data that includes position of QTL and outcome. Adds ticks to the figure.}

\item{expland_lim_factor}{DEFAULT: 0.025. Scale to increase plot limits by.}
}
\description{
This function takes the posterior odds results from bmediatR() and plots the genome-wide scan.
}
\examples{
plot_posterior_odds()
}
