% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediation_bf.R
\name{plot_posterior_stack}
\alias{plot_posterior_stack}
\title{Stacked barplot of posterior model probabilities function}
\usage{
plot_posterior_stack(med_bf_object, med_annot, protein.id)
}
\description{
This function takes posterior probability results from mediation_bf() and plots the stacked barplot of posterior model probabilities.
}
\examples{
plot_posterior_stack()
}
