% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{sim_balanced_locus}
\alias{sim_balanced_locus}
\title{Simulate a balanced eight allele design matrix}
\usage{
sim_balanced_locus(founder_allele_reps = 10)
}
\arguments{
\item{founder_allele_reps}{DEFAULT: 10. The number of observations of each founder allele.
For example, the default of 10 will produce a design matrix representing 80 artificial CC
strains.}
}
\description{
This function simulates a balanced eight allele design 
matrix for a specified number of observations of each
founder allele.
}
\examples{
sim_balanced_locus()
}
