% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{sim_cc_data}
\alias{sim_cc_data}
\title{Simulate Collaborative Cross (CC) data from realized CC genomes}
\usage{
sim_cc_data(
  cc_genoprobs,
  cc_map,
  qtl_effect_size = 0.1,
  strain_effect_size = 0,
  locus = NULL,
  vary_locus = TRUE,
  num_replicates = 1,
  num_sim = 1,
  M_ID = NULL,
  sample_method = c("uniform", "crp"),
  num_alleles = 8,
  num_founders = 8,
  beta = NULL,
  impute = TRUE,
  sim_label = "sim_m",
  return_means = TRUE
)
}
\arguments{
\item{cc_genoprobs}{A qtl2 formatted genoprobs object for CC mice.}

\item{cc_map}{A qtl2 formatted map object that corresponds to the genoprobs.}

\item{qtl_effect_size}{DEFAULT = 0.1. The proportion of variation in the simulated outcomes
explained by the QTL.}

\item{num_replicates}{DEFAULT: 1. The number of strain replicates to use.}

\item{num_sim}{DEFAULT: 1. The number of simulated outcomes to produce.}
}
\description{
This function takes qtl2 formatted genoprobs of CC mice and simulates outcomes/mediators driven by QTL.
}
\examples{
sim_cc_data()
}
