% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{sim_mpp_data}
\alias{sim_mpp_data}
\title{Simulate multiparental population (MPP) data from realized genomes}
\usage{
sim_mpp_data(
  genoprobs,
  map,
  qtl_effect_size = 0.1,
  strain_effect_size = 0,
  locus = NULL,
  vary_locus = TRUE,
  num_replicates = 1,
  num_sim = 1,
  M_ID = NULL,
  sample_method = c("uniform", "crp"),
  num_alleles = 8,
  num_founders = 8,
  beta = NULL,
  impute = TRUE,
  sim_label = "sim_m",
  return_means = TRUE
)
}
\arguments{
\item{genoprobs}{A qtl2 formatted genoprobs object from qtl2.}

\item{map}{A qtl2 formatted map object that corresponds to the genoprobs.}

\item{qtl_effect_size}{DEFAULT = 0.1. The proportion of variation in the simulated outcomes
explained by the QTL.}

\item{num_replicates}{DEFAULT: 1. The number of strain replicates to use.}

\item{num_sim}{DEFAULT: 1. The number of simulated outcomes to produce.}
}
\description{
This function takes qtl2 formatted genoprobs from an MPP population (e.g., CC or DO) and simulates 
outcomes driven by QTL.
}
\examples{
sim_mpp_data()
}
