% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{sim_mpp_single_locus}
\alias{sim_mpp_single_locus}
\title{Simulate multiparental population (MPP) data from a specified locus design matrix}
\usage{
sim_mpp_single_locus(
  locus_matrix,
  num_replicates,
  num_sim,
  qtl_effect_size = 0.1,
  strain_effect_size = 0,
  M_ID = NULL,
  sample_method = c("uniform", "crp"),
  num_alleles = 8,
  num_founders = 8,
  beta = NULL,
  impute = TRUE,
  sim_label = "sim_m",
  return_means = TRUE
)
}
\arguments{
\item{locus_matrix}{A design matrix of founder haplotypes at QTL for MPP founder strains. The number of 
rows should correspond to the number of strains. The number of columns is eight for the CC/DO.}

\item{num_replicates}{DEFAULT: 1. The number of strain replicates to use.}

\item{num_sim}{DEFAULT: 1. The number of simulated outcomes to produce.}

\item{qtl_effect_size}{DEFAULT = 0.1. The proportion of variation in the simulated outcomes
explained by the QTL.}
}
\description{
This function takes a locus design matrix, e.g., founder haplotypes dosages from CC mice,
to simulate outcomes driven by its QTL.
}
\examples{
sim_cc_single_locus()
}
