% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mediation.R
\name{sim_target_from_mediator}
\alias{sim_target_from_mediator}
\title{Simulate target from mediator}
\usage{
sim_target_from_mediator(M, mediator_effect_size = 0.5, sim_label = "sim_y")
}
\arguments{
\item{M}{Vector or matrix of mediator variables. Multiple mediator variables are supported. 
Names or rownames must match across y and X, and Z and w (if provided).}

\item{mediator_effect_size}{DEFAULT = 0.5. The proportion of variation in the simulated outcomes
explained by the mediator.}
}
\description{
This function takes a mediator matrix and simulates downstream
outcomes or targets.
}
\examples{
sim_target_from_mediator()
}
