% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmediatR.R
\name{bmediatR_v0}
\alias{bmediatR_v0}
\title{Bayesian model selection for mediation analysis function, version 0}
\usage{
bmediatR_v0(
  y,
  M,
  X,
  Z = NULL,
  w = NULL,
  kappa = rep(0.001, 8),
  lambda = rep(0.001, 8),
  tau_sq_mu = rep(1000, 8),
  tau_sq_Z = rep(1000, 8),
  phi_sq_X = c(NA, NA, 1, 0.5, NA, 1, NA, 0.5),
  phi_sq_m = c(NA, 1, NA, 0.5, NA, NA, NA, NA),
  phi_sq_y = c(NA, NA, NA, NA, NA, NA, 1, 0.5),
  ln_prior_c = "complete",
  options_X = list(sum_to_zero = TRUE, center = FALSE, scale = FALSE),
  verbose = T
)
}
\arguments{
\item{y}{Vector or single column matrix of an outcome variable. Single outcome variable expected. 
Names or rownames must match across M and X, and Z and w (if provided).}

\item{M}{Vector or matrix of mediator variables. Multiple mediator variables are supported. 
Names or rownames must match across y and X, and Z and w (if provided).}

\item{X}{Design matrix of the driver. Names or rownames must match across y and M, and Z and w (if provided).
One common application is for X to represent genetic information at a QTL, either as founder strain haplotypes
or variant genotypes, though X is generalizable to other types of variables.}

\item{Z}{DEFAULT: NULL. Design matrix of covariates that influence the outcome and mediator variables. 
Names or rownames must match across y, M and X, and w (if provided).}

\item{w}{DEFAULT: NULL. Vector or single column matrix of weights for individuals in analysis. 
Names must match across y, M X, and Z (if provided). A common use would be for an analysis of strain means, where w 
is a vector of the number of individuals per strain. If no w is given, observations are equally weighted as 1s.}

\item{tau_sq_mu}{DEFAULT: 1000. Variance component for the intercept. The DEFAULT represents a diffuse prior, analagous to 
a fixed effect term.}

\item{tau_sq_Z}{DEFAULT: 1000. Variance component for the covariates encoded in Z. The DEFAULT represents a diffuse prior, analagous 
to fixed effect terms.}

\item{ln_prior_c}{DEFAULT: "complete". The prior log case probabilities. See model_info() for description of likelihoods and their
combinations into cases. Simplified pre-set options are available, including "complete", "partial", and "reactive".}
}
\description{
This function takes an outcome (y), candidate mediators (M), and a driver as a design matrix (X) to perform a 
Bayesian model selection analysis for mediation. Version 0 has more flexibility for specifying the hyper priors
on the effect sizes, i.e., the variance explained by the explanatory variables at each step. It also does not
have the option for covariates and weights specific to M and y.
}
\examples{
bmediatR_v0()
}
