% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_mediation_bf.R
\name{get_perm_pval}
\alias{get_perm_pval}
\title{Function that calculate permutation-based p-values for Bayes factors}
\usage{
get_perm_pval(
  y,
  M,
  X,
  Z = NULL,
  num_perm = 1000,
  verbose = FALSE,
  bf_type = "lnBF_totmed",
  ...
)
}
\description{
This function takes Bayes factor results from mediation_bf() and calculates permutation p-values based
on permuting the mediator.
}
\examples{
get_perm_pval()
}
