% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediation_bf.R
\name{plot_posterior_bar}
\alias{plot_posterior_bar}
\title{Barplot of posterior model probabilities function}
\usage{
plot_posterior_bar(
  med_bf_object,
  med_annot = NULL,
  mediator_id,
  med_var = "protein.id",
  stack = FALSE
)
}
\description{
This function takes posterior probability results from mediation_bf() and plots barplots of posterior model probabilities.
}
\examples{
plot_posterior_bar()
}
