% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmediatR.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Summaries of posterior model probabilities function}
\usage{
posterior_summary(ln_prob_data, ln_prior_c, c_numerator)
}
\arguments{
\item{ln_prob_data}{Log posterior likelihoods under the various models, returned by bmediatR().}

\item{ln_prior_c}{Log prior model probabilities. If posterior_summary() is being used for a non-default posterior odds
summary, the log prior model probabilities used with bmediatR() are stored in its output.}

\item{c_numerator}{The index of models to be summed in the numberator of the posterior odds. Models and their order provided with
model_summary().}
}
\description{
This function takes the log posterior probability of the data (posterior likelihood) for the various models, the log prior model probabilities, and
returns log posterior odds
}
\examples{
posterior_summary()
}
